Remove focuspolicy (#230)
- Changed: The plugin no longer respects bevy_ui's `FocusPolicy`. This was proving to cause problems
  as mod_picking and bevy_ui have some fundamental differences that cannot be reconciled. This has
  been replaced by added fields on the `Pickable` component. You can use this to override the
  behavior of any entity in picking. This allows you to decide if that entity will block lower
  entities (on by default), and if that entity should emit events and be hover-able (on by default).
    - To make objects non-pickable, instead of removing the `Pickable` entity, use the new const
      value `Pickable::IGNORE`.
- Changed: The `Pickable` component is now optional. Backends can choose to use it for optimization,
  but should not filter out entities that do not have this component. An example of an optimization
  would be to early exit and stop raycasting once an entity is hit that blocks the entities below
  it.
- Changed: To fully remove bevy_ui as a dependency and avoid issues similar to the `FocusPolicy`
  change, bevy_mod_picking no longer updates the bevy_ui `Interaction` state. This has been replaced
  with a `PickInteraction` component that serves a similar purpose. This component aggregates the
  picking state of an entity (press, hover, none) across *all* pointers.
- Changed: `PickLayer`, used to order data from backends that targets the same render target, such
  as multiple render passes on the same window, has been changed from an `isize` to an `f32`. This
  change was made to support bevy_ui, which is "special" and can be rendered on any camera via a
  flag, instead of being rendered with its own camera. The bevy_ui backend now sets the order of any
  events emitted to be the camera order plus 0.5, which was not possible with an integer.
- Changed: The `PointerInteraction` component, which is added to pointers and tracks all entities
  being interacted with has changed internally from a hashmap of entities and their `Interaction` to
  a sorted list of entities and `HitData`, sorted by depth. This better reflects how pointer data
  would be expected to work, and makes it easier to find the topmost entity under each pointer.
- Added: `get_nearest_hit` method added to `PointerInteraction`
- Changed: Moved `PointerInteraction` from the `focus` module to `pointer`.

+++ crates/bevy_picking_core/src/backend.rs
@@ -17,7 +17,9 @@
 //! - The [`PointerHits`] events produced by a backend do **not** need to be sorted or filtered, all
 //! that is needed is an unordered list of entities and their [`HitData`].
 //!
-//! - **Backends should only pick entities with the [`Pickable`](crate::Pickable) component.**
+//! - Backends do not need to consider the [`Pickable`](crate::Pickable) component, though they may
+//!   use it for optimization purposes. For example, a backend that traverses a spatial hierarchy
+//!   may want to early exit if it intersects entity that blocks lower entities from being picked.
 
 use bevy::prelude::*;
 
@@ -52,14 +54,21 @@ pub struct PointerHits {
     /// pointer, then sorted by order, and checked sequentially, sorting each `PointerHits` by
     /// entity depth. Events with a higher `order` are effectively on top of events with a lower
     /// order.
-    pub order: isize,
+    ///
+    /// ### Why is this an `f32`???
+    ///
+    /// Bevy UI is special in that it can share a camera with other things being rendered. in order
+    /// to properly sort them, we need a way to make bevy_ui's order a tiny bit higher, like adding
+    /// 0.5 to the order. We can't use integers, and we want users to be using camera.order by
+    /// default, so this is the best solution at the moment.
+    pub order: f32,
 }
 
 /// Holds data from a successful pointer hit test.
 ///
 /// `depth` only needs to be self-consistent with other [`PointerHits`]s using the same
 /// [`RenderTarget`](bevy::render::camera::RenderTarget).
-#[derive(Clone, Copy, Debug, PartialEq, Reflect)]
+#[derive(Clone, Debug, PartialEq, Reflect)]
 pub struct HitData {
     /// The camera entity used to detect this hit. Useful when you need to find the ray that was
     /// casted for this hit when using a raycasting backend.
